new dialgo to delete

+++ Client/src/app/main/edit-profile/edit-profile.component.ts
@@ -1,8 +1,12 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from "@angular/router";
 import {environment} from "../../../environments/environment"
 import {HttpHelperService} from "../../services/http-helper.service";
 import {AuthHelperService} from "../../services/auth-helper.service";
+import { MatDialog } from "@angular/material/dialog";
+import {DeleteConfirmComponent} from "../../dialog/delete-confirm/delete-confirm.component";
 import { User } from 'src/app/types/user';
+
 
 
 @Component({
@@ -24,7 +28,9 @@ export class EditProfileComponent implements OnInit{
   }
 
 
-  constructor(public authService: AuthHelperService,
+  constructor(public router: Router,
+              public dialog: MatDialog,
+              public authService: AuthHelperService,
               private http: HttpHelperService,
               private auth: AuthHelperService) { }
 
@@ -40,8 +46,25 @@ export class EditProfileComponent implements OnInit{
   }
 
 
-  Delete(){
-    this.http.delete(environment.backend+"/users/" + this.auth.user!.username)
+  Delete() : void{
+    this.dialog.open(DeleteConfirmComponent,{
+      data: '¿Deseas borrar la cuenta?'
+    })
+      .afterClosed()
+      .subscribe((confirm: boolean) =>{
+        if(confirm){
+          this.http.delete(environment.backend+"/users/" + this.auth.user!.username)
+            .subscribe()
+          this.auth.logout();
+          this.router.navigate(['/login'])
+
+
+        }
+
+      }
+
+        );
+
   }
